Fixes #3084: Some refactoring to make better use of environment variables in the tests
Also, now expects OpenStack Identity V3 service to be of type 'identityv3'
