+++ lib/fog/openstack/identity.rb
@@ -1,4 +1,6 @@
 require 'fog/openstack/core'
+require 'fog/openstack/identity_v2'
+require 'fog/openstack/identity_v3'
 
 module Fog
   module Identity
@@ -19,48 +21,6 @@ module Fog
           service = Fog::Service.new(args)
         end
         service
-      end
-
-      module Common
-
-        def authenticate
-          if !@openstack_management_url || @openstack_must_reauthenticate
-            options = {
-                :openstack_api_key => @openstack_api_key,
-                :openstack_username => @openstack_username,
-                :openstack_auth_token => @openstack_must_reauthenticate ? nil : @openstack_auth_token,
-                :openstack_auth_uri => @openstack_auth_uri,
-                :openstack_tenant => @openstack_tenant,
-                :openstack_service_type => @openstack_service_type,
-                :openstack_service_name => @openstack_service_name,
-                :openstack_endpoint_type => @openstack_endpoint_type,
-                :openstack_region => @openstack_region
-            }
-
-            credentials = Fog::OpenStack.authenticate(options, @connection_options)
-
-            @current_user = credentials[:user]
-            @current_tenant = credentials[:tenant]
-
-            @openstack_must_reauthenticate = false
-            @auth_token = credentials[:token]
-            @openstack_management_url = credentials[:server_management_url]
-            @openstack_current_user_id = credentials[:current_user_id]
-            @unscoped_token = credentials[:unscoped_token]
-            uri = URI.parse(@openstack_management_url)
-          else
-            @auth_token = @openstack_auth_token
-            uri = URI.parse(@openstack_management_url)
-          end
-
-          @host = uri.host
-          @path = uri.path
-          @path.sub!(/\/$/, '')
-          @port = uri.port
-          @scheme = uri.scheme
-          true
-        end
-
       end
     end
