+++ spec/fog/openstack/identity_v3/idv3_service.yml
@@ -34,9 +34,9 @@ http_interactions:
       encoding: US-ASCII
       string: ! '{"services": [{"name": "keystone", "links": {"self": "http://devstack.openstack.stack:35357/v3/services/8e53f1d389df4059aeab1acfece2fc66"},
         "enabled": true, "type": "identity", "id": "8e53f1d389df4059aeab1acfece2fc66",
-        "description": "OpenStack Identity"}, {"name": "identity_v3", "links": {"self":
+        "description": "OpenStack Identity"}, {"name": "identityv3", "links": {"self":
         "http://devstack.openstack.stack:35357/v3/services/a47e0f0014fa409993cef0bd984a5ac6"},
-        "enabled": true, "type": "identity_v3", "id": "a47e0f0014fa409993cef0bd984a5ac6",
+        "enabled": true, "type": "identityv3", "id": "a47e0f0014fa409993cef0bd984a5ac6",
         "description": "OpenStack KeyStone V3 Service"}, {"description": "Test for
         endpoint parameter expansion; can be deleted", "links": {"self": "http://devstack.openstack.stack:35357/v3/services/d36a1f72832b428688b7dae7d0ab8a39"},
         "enabled": true, "type": "test_stefan", "id": "d36a1f72832b428688b7dae7d0ab8a39",
@@ -80,9 +80,9 @@ http_interactions:
       encoding: US-ASCII
       string: ! '{"services": [{"name": "keystone", "links": {"self": "http://devstack.openstack.stack:35357/v3/services/8e53f1d389df4059aeab1acfece2fc66"},
         "enabled": true, "type": "identity", "id": "8e53f1d389df4059aeab1acfece2fc66",
-        "description": "OpenStack Identity"}, {"name": "identity_v3", "links": {"self":
+        "description": "OpenStack Identity"}, {"name": "identityv3", "links": {"self":
         "http://devstack.openstack.stack:35357/v3/services/a47e0f0014fa409993cef0bd984a5ac6"},
-        "enabled": true, "type": "identity_v3", "id": "a47e0f0014fa409993cef0bd984a5ac6",
+        "enabled": true, "type": "identityv3", "id": "a47e0f0014fa409993cef0bd984a5ac6",
         "description": "OpenStack KeyStone V3 Service"}, {"description": "Test for
         endpoint parameter expansion; can be deleted", "links": {"self": "http://devstack.openstack.stack:35357/v3/services/d36a1f72832b428688b7dae7d0ab8a39"},
         "enabled": true, "type": "test_stefan", "id": "d36a1f72832b428688b7dae7d0ab8a39",
@@ -126,9 +126,9 @@ http_interactions:
       encoding: US-ASCII
       string: ! '{"services": [{"name": "keystone", "links": {"self": "http://devstack.openstack.stack:35357/v3/services/8e53f1d389df4059aeab1acfece2fc66"},
         "enabled": true, "type": "identity", "id": "8e53f1d389df4059aeab1acfece2fc66",
-        "description": "OpenStack Identity"}, {"name": "identity_v3", "links": {"self":
+        "description": "OpenStack Identity"}, {"name": "identityv3", "links": {"self":
         "http://devstack.openstack.stack:35357/v3/services/a47e0f0014fa409993cef0bd984a5ac6"},
-        "enabled": true, "type": "identity_v3", "id": "a47e0f0014fa409993cef0bd984a5ac6",
+        "enabled": true, "type": "identityv3", "id": "a47e0f0014fa409993cef0bd984a5ac6",
         "description": "OpenStack KeyStone V3 Service"}, {"description": "Test for
         endpoint parameter expansion; can be deleted", "links": {"self": "http://devstack.openstack.stack:35357/v3/services/d36a1f72832b428688b7dae7d0ab8a39"},
         "enabled": true, "type": "test_stefan", "id": "d36a1f72832b428688b7dae7d0ab8a39",
@@ -172,9 +172,9 @@ http_interactions:
       encoding: US-ASCII
       string: ! '{"services": [{"name": "keystone", "links": {"self": "http://devstack.openstack.stack:35357/v3/services/8e53f1d389df4059aeab1acfece2fc66"},
         "enabled": true, "type": "identity", "id": "8e53f1d389df4059aeab1acfece2fc66",
-        "description": "OpenStack Identity"}, {"name": "identity_v3", "links": {"self":
+        "description": "OpenStack Identity"}, {"name": "identityv3", "links": {"self":
         "http://devstack.openstack.stack:35357/v3/services/a47e0f0014fa409993cef0bd984a5ac6"},
-        "enabled": true, "type": "identity_v3", "id": "a47e0f0014fa409993cef0bd984a5ac6",
+        "enabled": true, "type": "identityv3", "id": "a47e0f0014fa409993cef0bd984a5ac6",
         "description": "OpenStack KeyStone V3 Service"}, {"description": "Test for
         endpoint parameter expansion; can be deleted", "links": {"self": "http://devstack.openstack.stack:35357/v3/services/d36a1f72832b428688b7dae7d0ab8a39"},
         "enabled": true, "type": "test_stefan", "id": "d36a1f72832b428688b7dae7d0ab8a39",
